Refactor config, refactor States, make State ovverides simpler, add State to Sprite map

+++ src/lib/store/States/Enemy/Die.svelte.ts
@@ -1,10 +1,6 @@
-/**
- * EnemyStates.ts
- * State classes for enemy behavior
- */
-
-import { BaseState } from '$lib/store/States/BaseState.svelte';
 import { managers } from '$store/managers.svelte';
+import { BaseState } from '../BaseState.svelte';
+
 export class Die extends BaseState {
 	constructor(stateMachine) {
 		super(stateMachine);
@@ -13,19 +9,14 @@ export class Die extends BaseState {
 	}
 
 	update() {
-		this.entity.rotation += 10;
-
-		if (this.entity.sprite.isAnimationComplete) {
+		if (this.entity.sprite && this.entity.sprite.isAnimationComplete) {
 			const { entityManager, stageManager } = managers.get(['entityManager', 'stageManager']);
 
+			stageManager.spawnEntity('loot', this.entity.position.clone(), {
+				target: entityManager.throne,
+				spawner: this.entity
+			});
 			entityManager.destroy(this.entity.id);
-
-			if (1) {
-				stageManager.spawnEntity('loot', this.entity.position.clone(), {
-					target: entityManager.throne,
-					spawner: this.entity
-				});
-			}
 		}
 	}
 }
